[PB-2411] refactor add member
